fix: 🐛 update configuration
webpack configuration is now under his own package ! maybe it shoudl be
published ?
