--- configurations/typescript/tsconfig.base.json
@@ -1,26 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "ESNext",
-    "module": "ES2015",
-    "lib": ["dom", "dom.iterable", "esnext"],
-    "skipLibCheck": true,
-    "allowJs": true,
-    "checkJs": true,
-    "removeComments": true,
-    "importHelpers": true,
-    "downlevelIteration": true,
-    "isolatedModules": true,
-    "strict": true,
-    "noUnusedLocals": true,
-    "resolveJsonModule": true,
-    "noUnusedParameters": true,
-    "pretty": true,
-    "noImplicitReturns": true,
-    "moduleResolution": "node",
-    "typeRoots": ["../../node_modules/@types"],
-    "types": ["node", "jest", "react"],
-    "allowSyntheticDefaultImports": true,
-    "esModuleInterop": true,
-    "forceConsistentCasingInFileNames": true
-  }
-}
