--- configurations/typescript/tsconfig.json
@@ -1,10 +1,0 @@
-{
-  "extends": "./tsconfig.base.json",
-  "compilerOptions": {
-    "target": "es5",
-    "module": "esnext",
-    "noImplicitAny": true,
-    "composite": true,
-    "jsx": "react",
-  }
-}
