--- configurations/typescript/tsconfig.react.json
@@ -1,14 +1,0 @@
-{
-  "compilerOptions": {
-    "target": "esnext",
-    "module": "commonjs",
-    "lib": ["dom", "dom.iterable", "esnext", "es6", "es2016.array.include", "es2017.object"],
-    "isolatedModules": true,
-    "strict": true,
-    "moduleResolution": "node",
-    "jsx": "react",
-    "allowSyntheticDefaultImports": true,
-    "esModuleInterop": true,
-  },
-  "exclude": ["node_modules"]
-}
