--- configurations/webpack/plugins/typescript.js
@@ -1,5 +1,0 @@
-const ForkTsCheckerWebpackPlugin = require('fork-ts-checker-webpack-plugin');
-
-module.exports = {
-  views: [new ForkTsCheckerWebpackPlugin()],
-};
