+++ package.json
@@ -35,64 +35,28 @@
     "test": "cross-env jest --passWithNoTests --coverage --reporters=default --reporters=jest-junit"
   },
   "devDependencies": {
-    "@babel/core": "^7.10.5",
-    "@babel/plugin-proposal-class-properties": "^7.10.4",
-    "@babel/plugin-proposal-private-methods": "^7.10.4",
-    "@babel/preset-env": "^7.10.4",
-    "@babel/preset-react": "^7.10.4",
-    "@microsoft/webpack-project-references-alias": "^1.1.1",
-    "@testing-library/react": "^11.0.1",
-    "@testing-library/react-hooks": "^3.4.1",
     "@types/jest": "^26.0.7",
     "@types/node": "^14.0.19",
     "@typescript-eslint/eslint-plugin": "^4.1.1",
     "@typescript-eslint/parser": "^4.1.1",
-    "@zerollup/ts-transform-paths": "^1.7.18",
-    "antd-dayjs-webpack-plugin": "^1.0.0",
-    "babel-loader": "^8.1.0",
-    "babel-plugin-import": "^1.13.0",
-    "bundle-stats": "^2.2.0",
-    "bundle-stats-webpack-plugin": "^2.2.0",
-    "circular-dependency-plugin": "^5.2.0",
     "conventional-changelog-cli": "^2.0.34",
     "cross-env": "^7.0.2",
-    "css-loader": "^4.0.0",
     "eslint": "^7.5.0",
     "eslint-config-prettier": "^6.11.0",
     "eslint-plugin-prettier": "^3.1.4",
     "eslint-plugin-react": "^7.20.3",
     "eslint-plugin-react-hooks": "^4.0.8",
-    "file-loader": "^6.0.0",
-    "fork-ts-checker-webpack-plugin": "^5.0.7",
     "git-cz": "^4.7.0",
-    "html-webpack-plugin": "^4.3.0",
     "husky": "^4.2.5",
     "hygen": "^6.0.4",
-    "identity-obj-proxy": "^3.0.0",
-    "inspectpack": "^4.5.2",
     "jest": "^26.1.0",
     "jest-junit": "^11.1.0",
     "lerna": "^3.22.1",
     "lerna-changelog": "^1.0.1",
-    "less-loader": "^7.0.1",
     "lint-staged": "^10.2.11",
-    "mini-css-extract-plugin": "^0.11.2",
     "prettier": "^2.0.5",
-    "react-test-renderer": "^16.13.1",
     "rimraf": "^3.0.2",
-    "sass-loader": "^10.0.2",
-    "source-map-loader": "^1.0.1",
-    "style-loader": "^1.2.1",
-    "ts-import-plugin": "^1.6.6",
-    "ts-jest": "^26.1.1",
-    "ts-loader": "^8.0.4",
     "typedoc": "^0.19.2",
-    "typescript": "^4.0.2",
-    "url-loader": "^4.1.0",
-    "webpack": "^4.43.0",
-    "webpack-bundle-analyzer": "^3.8.0",
-    "webpack-cli": "^3.3.12",
-    "webpack-dev-server": "^3.11.0",
-    "webpack-node-externals": "^2.5.2"
+    "typescript": "^4.0.2"
   }
 }
