+++ packages/pack/README.md
@@ -1,0 +1,11 @@
+# @organization/pack
+
+> TODO: description
+
+## Usage
+
+```
+const pack = require('@organization/pack');
+
+// TODO: DEMONSTRATE API
+```
