+++ packages/pack/package.json
@@ -1,0 +1,65 @@
+{
+  "name": "@organization/pack",
+  "version": "0.0.1",
+  "description": "YOUR DESCRIPTION HERE",
+  "main": "dist/index.js",
+  "types": "dist/index.d.ts",
+  "license": "MIT",
+  "private": true,
+  "files": [
+    "dist"
+  ],
+  "scripts": {
+    "build:watch": "tsdx build --watch",
+    "build": "yarn prebuild && tsdx build",
+    "clean": "cross-env rimraf ./dist",
+    "docs": "typedoc",
+    "prebuild": "yarn clean"
+  },
+  "dependencies": {
+    "@babel/core": "7.12.3",
+    "@babel/plugin-proposal-class-properties": "7.12.1",
+    "@babel/plugin-proposal-private-methods": "7.12.1",
+    "@babel/plugin-syntax-object-rest-spread": "7.8.3",
+    "@babel/plugin-transform-arrow-functions": "7.12.1",
+    "@babel/plugin-transform-regenerator": "7.12.1",
+    "@babel/plugin-transform-runtime": "7.12.1",
+    "@babel/preset-env": "7.12.1",
+    "@babel/preset-react": "7.12.5",
+    "@babel/preset-typescript": "7.12.7",
+    "@babel/runtime-corejs3": "7.12.5",
+    "@microsoft/webpack-project-references-alias": "1.1.1",
+    "@zerollup/ts-transform-paths": "1.7.18",
+    "babel-loader": "8.2.1",
+    "bundle-stats": "2.3.0",
+    "bundle-stats-webpack-plugin": "2.3.0",
+    "circular-dependency-plugin": "5.2.2",
+    "core-js": "3.8.1",
+    "css-loader": "5.0.1",
+    "file-loader": "6.2.0",
+    "html-webpack-plugin": "4.5.0",
+    "inspectpack": "4.5.2",
+    "less-loader": "7.1.0",
+    "mini-css-extract-plugin": "1.3.1",
+    "optimize-css-assets-webpack-plugin": "5.0.4",
+    "sass": "1.29.0",
+    "sass-loader": "10.1.0",
+    "source-map-loader": "1.1.2",
+    "style-loader": "2.0.0",
+    "ts-import-plugin": "1.6.6",
+    "ts-loader": "8.0.11",
+    "url-loader": "4.1.1",
+    "webpack": "5.5.1",
+    "webpack-bundle-analyzer": "4.1.0",
+    "webpack-cli": "4.2.0",
+    "webpack-dev-server": "3.11.0",
+    "webpack-merge": "5.4.0",
+    "webpack-node-externals": "2.5.2"
+  },
+  "devDependencies": {
+    "@types/mini-css-extract-plugin": "1.2.1",
+    "@types/webpack": "4.41.25",
+    "@types/webpack-bundle-analyzer": "3.9.0",
+    "tsdx": "0.14.1"
+  }
+}
